# okhantsev.vladimir/project/backend/internal/server/server.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
)

// Defines values for CreateOrderRequestDeliveryType.
const (
	Delivery CreateOrderRequestDeliveryType = "delivery"
	Pickup   CreateOrderRequestDeliveryType = "pickup"
)

// Valid indicates whether the value is a known member of the CreateOrderRequestDeliveryType enum.
func (e CreateOrderRequestDeliveryType) Valid() bool {
	switch e {
	case Delivery:
		return true
	case Pickup:
		return true
	default:
		return false
	}
}

// Defines values for CreateOrderRequestPaymentType.
const (
	Card CreateOrderRequestPaymentType = "card"
	Cash CreateOrderRequestPaymentType = "cash"
)

// Valid indicates whether the value is a known member of the CreateOrderRequestPaymentType enum.
func (e CreateOrderRequestPaymentType) Valid() bool {
	switch e {
	case Card:
		return true
	case Cash:
		return true
	default:
		return false
	}
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	Address      *string                        `json:"address,omitempty"`
	DeliveryType CreateOrderRequestDeliveryType `json:"deliveryType"`
	Email        *string                        `json:"email,omitempty"`
	Items        []OrderItem                    `json:"items"`
	NeedWrapping *bool                          `json:"needWrapping,omitempty"`
	PaymentType  CreateOrderRequestPaymentType  `json:"paymentType"`
	Phone        string                         `json:"phone"`
}

// CreateOrderRequestDeliveryType defines model for CreateOrderRequest.DeliveryType.
type CreateOrderRequestDeliveryType string

// CreateOrderRequestPaymentType defines model for CreateOrderRequest.PaymentType.
type CreateOrderRequestPaymentType string

// Error defines model for Error.
type Error struct {
	Message *string `json:"message,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// Order defines model for Order.
type Order struct {
	Date       time.Time   `json:"date"`
	Id         int         `json:"id"`
	Items      []OrderItem `json:"items"`
	ItemsCount int         `json:"itemsCount"`
	TotalPrice int         `json:"totalPrice"`
}

// OrderItem defines model for OrderItem.
type OrderItem struct {
	GoodId   int `json:"goodId"`
	Quantity int `json:"quantity"`
}

// Product defines model for Product.
type Product struct {
	Characteristics *map[string]string `json:"characteristics,omitempty"`
	Colors          *[]string          `json:"colors,omitempty"`
	Description     *string            `json:"description,omitempty"`
	Id              int                `json:"id"`
	IsHit           *bool              `json:"isHit,omitempty"`
	IsNew           *bool              `json:"isNew,omitempty"`
	Name            string             `json:"name"`
	Photo           *string            `json:"photo,omitempty"`
	Popularity      *int               `json:"popularity,omitempty"`
	Price           int                `json:"price"`
	Rating          float32            `json:"rating"`
	Type            *string            `json:"type,omitempty"`
}

// User defines model for User.
type User struct {
	Username string `json:"username"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetGoods List goods
	// (GET /goods)
	GetGoods(w http.ResponseWriter, r *http.Request)
	// Login Login
	// (POST /login)
	Login(w http.ResponseWriter, r *http.Request)
	// Logout Logout
	// (POST /logout)
	Logout(w http.ResponseWriter, r *http.Request)
	// GetOrders Get order history
	// (GET /orders)
	GetOrders(w http.ResponseWriter, r *http.Request)
	// CreateOrder Create order
	// (POST /orders)
	CreateOrder(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// GetGoods List goods
// (GET /goods)
func (_ Unimplemented) GetGoods(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Login Login
// (POST /login)
func (_ Unimplemented) Login(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Logout Logout
// (POST /logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetOrders Get order history
// (GET /orders)
func (_ Unimplemented) GetOrders(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreateOrder Create order
// (POST /orders)
func (_ Unimplemented) CreateOrder(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetGoods operation middleware
func (siw *ServerInterfaceWrapper) GetGoods(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetGoods(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOrders operation middleware
func (siw *ServerInterfaceWrapper) GetOrders(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOrders(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateOrder operation middleware
func (siw *ServerInterfaceWrapper) CreateOrder(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateOrder(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/login", wrapper.Login)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/goods", wrapper.GetGoods)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/orders", wrapper.GetOrders)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/orders", wrapper.CreateOrder)
	})

	return r
}