# okhantsev.vladimir/project/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Gadget Hub
  version: 1.0.0
paths:
  /login:
    post:
      summary: Login
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /logout:
    post:
      summary: Logout
      operationId: logout
      tags: [Auth]
      responses:
        '200':
          description: Session ended
  /goods:
    get:
      summary: List goods
      operationId: getGoods
      tags: [Goods]
      responses:
        '200':
          description: Array of goods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /orders:
    get:
      summary: Get order history
      operationId: getOrders
      tags: [Orders]
      responses:
        '200':
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Create order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    User:
      type: object
      required: [username]
      properties:
        username:
          type: string
    Product:
      type: object
      required: [id, name, price, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        rating:
          type: number
        popularity:
          type: integer
        photo:
          type: string
        description:
          type: string
        characteristics:
          type: object
          additionalProperties:
            type: string
        type:
          type: string
        colors:
          type: array
          items:
            type: string
        isNew:
          type: boolean
        isHit:
          type: boolean
    OrderItem:
      type: object
      required: [goodId, quantity]
      properties:
        goodId:
          type: integer
        quantity:
          type: integer
    CreateOrderRequest:
      type: object
      required: [phone, deliveryType, paymentType, items]
      properties:
        email:
          type: string
        phone:
          type: string
        deliveryType:
          type: string
          enum: [pickup, delivery]
        address:
          type: string
        paymentType:
          type: string
          enum: [card, cash]
        needWrapping:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    Order:
      type: object
      required: [id, date, items, totalPrice, itemsCount]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalPrice:
          type: integer
        itemsCount:
          type: integer
    Error:
      type: object
      properties:
        message:
          type: string